chore: add mock-disabled environment
