+++ apps/dh/app-dh/project.json
@@ -103,12 +103,6 @@
             "inputmask",
             "flat",
             "msw"
-          ],
-          "fileReplacements": [
-            {
-              "replace": "libs/dh/core/shell/src/lib/dh-core-shell.providers.ts",
-              "with": "libs/dh/core/shell/src/lib/dh-core-shell.providers-skip-auth.ts"
-            }
           ]
         },
         "mocked": {
@@ -120,10 +114,27 @@
             {
               "replace": "libs/dh/shared/environments/src/lib/environment.ts",
               "with": "libs/dh/shared/environments/src/lib/environment.mocked.ts"
-            },
+            }
+          ],
+          "allowedCommonJsDependencies": [
+            "date-fns",
+            "@mswjs/cookies",
+            "@mswjs/interceptors",
+            "msw",
+            "inputmask",
+            "flat",
+            "crypto-js"
+          ]
+        },
+        "auth-disabled": {
+          "optimization": false,
+          "extractLicenses": false,
+          "sourceMap": true,
+          "namedChunks": true,
+          "fileReplacements": [
             {
-              "replace": "libs/dh/core/shell/src/lib/dh-core-shell.providers.ts",
-              "with": "libs/dh/core/shell/src/lib/dh-core-shell.providers-skip-auth.ts"
+              "replace": "libs/dh/shared/environments/src/lib/environment.ts",
+              "with": "libs/dh/shared/environments/src/lib/environment.auth-disabled.ts"
             }
           ],
           "allowedCommonJsDependencies": [
@@ -161,6 +172,14 @@
           "sslCert": "localhost.crt",
           "host": "0.0.0.0",
           "buildTarget": "app-dh:build:mocked"
+        },
+        "auth-disabled": {
+          "hmr": true,
+          "ssl": true,
+          "sslKey": "localhost.key",
+          "sslCert": "localhost.crt",
+          "host": "0.0.0.0",
+          "buildTarget": "app-dh:build:auth-disabled"
         }
       },
       "defaultConfiguration": "mocked"
