+++ libs/dh/shared/environments/src/lib/dh-environment.ts
@@ -16,4 +16,5 @@
  */
 export interface DhEnvironment {
   readonly production: boolean;
+  readonly authDisabled: boolean;
 }
