+++ libs/dh/shared/environments/src/lib/environment.mocked.ts
@@ -45,4 +45,5 @@ import 'zone.js/plugins/zone-error';
  */
 export const environment: DhEnvironment = {
   production: false,
+  authDisabled: false,
 };
