+++ libs/dh/shared/environments/src/lib/environment.prod.ts
@@ -18,4 +18,5 @@ import { DhEnvironment } from './dh-environment';
 
 export const environment: DhEnvironment = {
   production: true,
+  authDisabled: false,
 };
